requirements.txt
colorama
matplotlib
pandas
requests
selenium
beautifulsoup4
pyserial